pyproject.toml
[tool.poetry]
name = "gawaipos"
version = "0.1.0"
description = "Admin Gawaipos"
authors = ["pbutarbutar <paru.liano@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
django-environment = "^0.3.8"
django-environ = "^0.8.1"
mysqlclient = "^2.1.0"
django-import-export = "^2.8.0"
django-money = "^2.1.1"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
